backend/requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.6
ecdsa==0.19.1
fastapi==0.103.2
greenlet==3.2.4
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
mysql-connector-python==8.0.33
passlib==1.7.4
pillow==11.3.0
protobuf==3.20.3
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.13
pydantic_core==2.33.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.53
sqlmodel==0.0.14
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.22.0
